switched to using ${applicationId} in manifest files; use BuildConfig.APPLICATION_ID in PushNotificationReceiver for CANCEL_PUSH Flavors implement App.Flavor for generating LayerClient, ParticipantProvider, and AuthenticationProvider.  Renamed flavors to order them better in AndroidStudio.  Renamed flavor classes from App* to Rails* and Demo* to differentiate them.
