--- app/src/herokuprovider/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<manifest
-    package="com.layer.messenger"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application>
-        <activity
-            android:name="com.layer.messenger.flavor.AppLoginActivity"
-            android:windowSoftInputMode="stateAlwaysVisible|adjustResize"/>
-    </application>
-</manifest>
