--- app/src/herokuprovider/res/values/strings.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<resources>
-    <string name="app_name">Atlas Messenger</string>
-
-    <string name="login_email_hint">Email</string>
-    <string name="login_password_hint">Password</string>
-    <string name="login_dialog_message">Logging In&#8230;</string>
-</resources>
