+++ app/src/main/res/layout/activity_app_settings.xml
@@ -191,29 +191,6 @@
             <TextView
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="@string/settings_title_android_version"
-                android:textAppearance="?android:attr/textAppearanceMedium"/>
-
-            <TextView
-                android:id="@+id/android_version"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="[android version]"
-                android:textAppearance="?android:attr/textAppearanceSmall"/>
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="72dp"
-            android:layout_gravity="center_vertical"
-            android:layout_marginLeft="16dp"
-            android:layout_marginRight="16dp"
-            android:orientation="vertical">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
                 android:text="@string/settings_title_atlas_version"
                 android:textAppearance="?android:attr/textAppearanceMedium"/>
 
@@ -245,6 +222,29 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:text="[layer version]"
+                android:textAppearance="?android:attr/textAppearanceSmall"/>
+
+        </LinearLayout>
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="72dp"
+            android:layout_gravity="center_vertical"
+            android:layout_marginLeft="16dp"
+            android:layout_marginRight="16dp"
+            android:orientation="vertical">
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="@string/settings_title_android_version"
+                android:textAppearance="?android:attr/textAppearanceMedium"/>
+
+            <TextView
+                android:id="@+id/android_version"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:text="[android version]"
                 android:textAppearance="?android:attr/textAppearanceSmall"/>
 
         </LinearLayout>
