+++ app/src/providerdemo/res/drawable/app_id_scanner_outline.xml
