+++ app/src/providerdemo/res/layout/activity_app_id_scanner.xml
