+++ app/src/providerrails/AndroidManifest.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest
+    package="com.layer.messenger"
+    xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <application>
+        <activity
+            android:name=".flavor.RailsLoginActivity"
+            android:windowSoftInputMode="stateAlwaysVisible|adjustResize"/>
+    </application>
+</manifest>
